fix(tui): correct metrics contract drift vs OMS backend (float QA/QC + partial-item re-fetch)
Cross-project validation (OMS metrics/cycle-count backend vs this client) found:
- quantity_available/quantity_committed are DRF FloatField (JSON 5.0), but the
  ItemMetrics struct declared them *int → Go could not unmarshal → the ENTIRE
  metrics fetch failed → the row silently never rendered against a real server.
  Changed both to *float64 + a metricFloatQtyString formatter (whole values drop
  the trailing .0).
- The cycle-count response is a PARTIAL item (id/current_stock/last_counted_at/
  days_since_last_count); the handler assigned it to s.item, blanking
  name/SKU/suppliers/costs after a count. Now re-fetches the full item + metrics.
- Hardened the metrics test fixture to emit float 5.0/0.0 (real DRF output) with
  explicit QA/QC decode assertions so this drift class is caught going forward.

go build/vet/test all green.

+++ internal/omsapi/inventory.go
@@ -174,8 +174,8 @@ func (c *Client) ScanItem(ctx context.Context, id string) (*Item, error) {
 type ItemMetrics struct {
 	CurrentStock      *int          `json:"current_stock"`       // QOH — quantity on hand
 	QuantityOnOrder   *int          `json:"quantity_on_order"`   // QOO — on open POs
-	QuantityAvailable *int          `json:"quantity_available"`  // QA  — on hand minus committed
-	QuantityCommitted *int          `json:"quantity_committed"`  // QC  — reserved
+	QuantityAvailable *float64      `json:"quantity_available"`  // QA  — on hand minus committed (backend FloatField)
+	QuantityCommitted *float64      `json:"quantity_committed"`  // QC  — reserved (backend FloatField)
 	QuantityInTransit *int          `json:"quantity_in_transit"` // QIT — shipped, not received
 	ReorderPoint      *int          `json:"reorder_point"`       // RP
 	LeadTimeDays      *float64      `json:"lead_time_days"`      // Lead — days (may be fractional avg)
